[Add] `Http::Request()` function, renamed HttpMsg.h -> http.h
