+++ source/server/userauth.h
@@ -20,7 +20,7 @@ along with Foobar. If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-#include "HttpMsg.h"
+#include "http.h"
 #include "UTFString.h"
 
 #include <vector>
